requirements.txt
# $ conda create --name <env> --file <this file>
# conda env export > environment.yml
# Dependencies from this file can be installed with the following command:
# I derived this file from the output of the following command and then edited it
# in GitHub Actions. 
# pip install -r requirements.txt
# platform: win-64
# Specific package versions are specified here to allow more consistent caching
# This file may be used to create an environment using:
# to match the appropriate syntax:
ABlog==0.11.11
black==24.8.0
chartbook[all]
colorama
doit==0.36.0
fabric==3.2.2
holidays
ipython
jaydebeapi
jupyter
jupyter-book
jupyterlab
linearmodels==6.1
linkify-it-py
matplotlib==3.9.2
myst-nb
myst-parser==2.0.0
notebook
numpy==1.26.4
numpydoc==1.8.0
openpyxl==3.1.5
pandas-datareader==0.10.0
pandas-market-calendars==4.4.1
pandas==2.2.3
paramiko==3.5.0
plotly==5.24.1
polars==1.9.0
pyarrow
pydata-sphinx-theme==0.15.4
pytest==8.3.3
python-decouple==3.8
python-dotenv==1.0.1
pyxlsb==1.0.10
requests==2.32.3
ruff
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
sphinx-autodoc2
sphinx-book-theme==1.1.3
sphinx-design
sphinx==7.3.7
sphinxext-opengraph
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxext-rediraffe
statsmodels==0.14.4
wrds==3.2.0
xbbg==0.7.7
xlrd==2.0.1
xlwings==0.33.3
zstandard==0.23.0